@@ -1,29 +1,25 @@
 {
   "name": "server",
   "version": "0.0.1",
-  "description": "Servidor",
-  "main": "server.ts",
+  "description": "Servidor para o projeto de hotelaria Accomodare Hotel",
+  "main": "server.js",
   "scripts": {
-    "test": "npm run tsc && jasmine --config=./spec/support/jasmine.json",
-    "tsc": "tsc",
-    "start": "npm run tsc && node typeScript/server.js"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "npx tsc",
+    "start": "node dist/server.js",
+    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/server.js\""
   },
-  "author": "Daniel Bastos",
-  "license": "MIT",
+  "author": "Joao Victor Cavalcanti",
+  "license": "ISC",
   "dependencies": {
-    "@types/body-parser": "^1.16.5",
-    "@types/express": "^4.0.37",
-    "body-parser": "^1.18.2",
-    "express": "^4.15.4",
-    "request": "^2.34",
-    "nodemailer": "^6.1.1"
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2"
   },
   "devDependencies": {
-    "@types/jasmine": "^2.8.9",
-    "@types/request-promise": "^4.1.38",
-    "jasmine": "^2.8.0",
-    "jasmine-promises": "^0.4.1",
-    "request-promise": "^4.2.2",
-    "typescript": "^3.3.3333"
+    "@types/express": "^4.17.17",
+    "@types/node": "^18.15.11",
+    "concurrently": "^8.0.1",
+    "nodemon": "^2.0.22",
+    "typescript": "^5.0.3"
   }
 }
